Update StockService to ensure numeric type for percentage fields and enhance stock search tests
- Modified the `StockService` to cast `period_change_percent` and `day_change_percent` to numeric during stock updates, ensuring data integrity.
- Refactored integration tests in `catalog-backfill.test.ts` to separately validate stock search results for 'COLD' and 'WARM', improving test clarity and coverage.

+++ packages/backend/src/modules/stock/stock.service.ts
@@ -256,8 +256,8 @@ export class StockService {
             const chunk = rows.slice(offset, offset + METRICS_REFRESH_CHUNK)
             await this.ctx.db.execute(sql`
                 UPDATE stock AS s SET
-                    period_change_percent = v.period_change_percent,
-                    day_change_percent = v.day_change_percent,
+                    period_change_percent = v.period_change_percent::numeric,
+                    day_change_percent = v.day_change_percent::numeric,
                     metrics_updated_at = ${metricsUpdatedAt}
                 FROM (VALUES ${sql.join(
                     chunk.map((row) => sql`(
